requirements.txt
# Core Polymarket client
py-clob-client>=0.18.0

# Environment and configuration
python-dotenv>=1.0.0

# Async HTTP and WebSocket
aiohttp>=3.9.0
websockets>=12.0
httpx>=0.25.0

# Web3 and blockchain
web3>=6.0.0
eth-account>=0.10.0

# Browser automation for Cloudflare bypass
undetected-chromedriver>=3.5.0
selenium>=4.15.0

# HTTP client
requests>=2.31.0
